Helpers refactor of helpers.py helpers :)
