+++ pytest_fixtures/component/rh_cloud.py
@@ -3,7 +3,6 @@ from broker import Broker
 
 from robottelo.config import settings
 from robottelo.constants import DEFAULT_SUBSCRIPTION_NAME
-from robottelo.helpers import file_downloader
 
 
 @pytest.fixture(scope='module')
@@ -19,9 +18,7 @@ def rhcloud_sat_host(satellite_factory):
 def rhcloud_manifest_org(rhcloud_sat_host):
     """A module level fixture to get organization with manifest."""
     org = rhcloud_sat_host.api.Organization().create()
-    manifests_path = file_downloader(
-        file_url=settings.fake_manifest.url['default'], hostname=rhcloud_sat_host.hostname
-    )[0]
+    manifests_path = rhcloud_sat_host.get(remote_path=settings.fake_manifest.url['default'])
     rhcloud_sat_host.cli.Subscription.upload({'file': manifests_path, 'organization-id': org.id})
     return org
