+++ robottelo/host_helpers/__init__.py
@@ -1,6 +1,7 @@
 from robottelo.host_helpers.capsule_mixins import CapsuleInfo
 from robottelo.host_helpers.contenthost_mixins import SystemFacts
 from robottelo.host_helpers.contenthost_mixins import VersionedContent
+from robottelo.host_helpers.host_mixins import IOHelpers
 from robottelo.host_helpers.satellite_mixins import ContentInfo
 from robottelo.host_helpers.satellite_mixins import Factories
 
@@ -14,4 +15,8 @@ class CapsuleMixins(CapsuleInfo):
 
 
 class SatelliteMixins(ContentInfo, Factories):
+    pass
+
+
+class HostMixins(IOHelpers):
     pass
