+++ tests/foreman/api/test_user.py
@@ -40,8 +40,8 @@ from robottelo.datafactory import parametrized
 from robottelo.datafactory import valid_data_list
 from robottelo.datafactory import valid_emails_list
 from robottelo.datafactory import valid_usernames_list
-from robottelo.helpers import gen_ssh_keypairs
 from robottelo.helpers import read_data_file
+from robottelo.utils import gen_ssh_keypairs
 
 
 @pytest.fixture(scope='module')
