+++ tests/foreman/ui/test_ldap_authentication.py
@@ -32,7 +32,6 @@ from robottelo.constants import CERT_PATH
 from robottelo.constants import LDAP_ATTR
 from robottelo.constants import PERMISSIONS
 from robottelo.datafactory import gen_string
-from robottelo.helpers import file_downloader
 from robottelo.rhsso_utils import delete_rhsso_group
 from robottelo.rhsso_utils import run_command
 
@@ -46,12 +45,7 @@ def set_certificate_in_satellite(server_type, target_sat, hostname=None):
     """update the cert settings in satellite based on type of ldap server"""
     if server_type == 'IPA':
         idm_cert_path_url = os.path.join(settings.ipa.hostname, 'ipa/config/ca.crt')
-        file_downloader(
-            file_url=idm_cert_path_url,
-            local_path=CERT_PATH,
-            file_name='ipa.crt',
-            hostname=target_sat.hostname,
-        )
+        target_sat.get(remote_path=idm_cert_path_url, local_path=CERT_PATH+'ipa.crt')
     elif server_type == 'AD':
         assert hostname is not None
         target_sat.execute('yum -y --disableplugin=foreman-protector install cifs-utils')
